passes tests + Wall

+++ shittyml/activation.cpp
@@ -2,7 +2,7 @@
 using namespace shittyml;
 
 vec ReLU::forward(vec a) {
-    for (int i = 0; i < a.size(); i++)
+    for (size_t i = 0; i < a.size(); i++)
         a[i] = a[i] > 0 ? a[i] : 0;
     return a;
 }
